login localstorage,netbank auto update,change password

+++ Banking_BackEnd/Banking/Models/Admin.cs
@@ -16,7 +16,7 @@ namespace Banking.Models
     {
         public int Admin_Id { get; set; }
         public string Password { get; set; }
-        public int Mobile_Number { get; set; }
+        public long Mobile_Number { get; set; }
         public string Email_Id { get; set; }
     }
 }
